Delete the `cmd2_ext_test.py` plugin code from the cmd2 repo
Also:
- Copy the ExternalTestMixin class implementation to the `tests_isolated_test_commandset/conftest.py` file
- Remove info on the External Test Plugin from the documentation
- Add info on the cmd2 Plugin Template to the documentation
- Remove loading of `cmd2-ext-test` source code from the pyproject.toml
- Removed ext-test stuff from tasks.py for invoke usage

There is likely a lot more cleanup that can and should be done. This is just a first attempt to get our GitHub Actions CI/CD working without this plugin's circular dependency.
