+++ docs/plugins/plugin_template.md
@@ -1,0 +1,7 @@
+# cmd2 Plugin Template
+
+## Overview
+
+The [cmd2 Plugin Template](https://github.com/python-cmd2/cmd2/tree/main/plugins/template) is a
+general example that shows you how you can develop a plugin for `cmd2`. Plugins are generally mixin
+classes that add some extra functionality to your class which inherits from [cmd2.Cmd][].
