--- plugins/ext_test/CHANGELOG.md
@@ -1,22 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project
-adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
-
-## 0.2.0 (2020-09-11)
-
-- Updated documentation to reflect new home inside of main cmd2 repo.
-- Updated python version requirements to match cmd2
-
-## 0.1.2 (2020-08-03)
-
-- Bug Fixes
-    - Applied fix to match change in cmd2 APIs
-
-## 0.1.1 (2020-03-09)
-
-### Added
-
-- Initial contribution
