--- plugins/ext_test/cmd2_ext_test/py.typed
@@ -1,1 +1,0 @@
-# PEP 561
