+++ pyproject.toml
@@ -66,7 +66,6 @@ docs = [
     "setuptools>=64",
     "setuptools_scm>=8",
 ]
-plugins = ["cmd2-ext-test"]
 quality = ["pre-commit>=2.20.0"]
 test = [
     "codecov>=2",
@@ -315,7 +314,4 @@ packages = ["cmd2"]
 [tool.setuptools_scm]
 
 [tool.uv]
-default-groups = ["build", "dev", "plugins"]
-
-[tool.uv.sources]
-cmd2-ext-test = { path = "plugins/ext_test", editable = true }
+default-groups = ["build", "dev"]
